@@ -11,11 +11,13 @@
 namespace Positibe\Bundle\NewsBundle\Controller;
 
 use Positibe\Bundle\NewsBundle\Entity\Post;
-use Positibe\Bundle\OrmMenuBundle\Model\MenuNodeReferrersInterface;
 use Sylius\Bundle\ResourceBundle\Controller\RequestConfiguration;
 use Sylius\Bundle\ResourceBundle\Controller\ResourceController;
+use Symfony\Cmf\Bundle\CoreBundle\Translatable\TranslatableInterface;
 use Symfony\Cmf\Bundle\SeoBundle\SeoAwareInterface;
+use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\Workflow\Exception\ExceptionInterface;
 
 
 /**
@@ -40,9 +42,13 @@ class PostController extends ResourceController
     public function findOr404(RequestConfiguration $configuration)
     {
         /** @var Post $post */
-        $post = parent::findOr404($configuration);;
+        $post = parent::findOr404($configuration);
 
-        if ($dataLocale = $configuration->getRequest()->get('data_locale')) {
+        if ($post instanceof TranslatableInterface && $dataLocale = $configuration->getRequest()->get(
+                'data_locale',
+                $this->getParameter('locale')
+            )
+        ) {
             $post->setLocale($dataLocale);
 
             if ($post instanceof SeoAwareInterface && $seoMetadata = $post->getSeoMetadata()) {
@@ -50,22 +56,46 @@ class PostController extends ResourceController
                 $this->get('doctrine.orm.entity_manager')->refresh($seoMetadata);
             }
 
-            if ($post instanceof MenuNodeReferrersInterface) {
-                foreach ($post->getMenuNodes() as $menu) {
-                    $menu->setLocale($dataLocale);
-                    $this->get('doctrine.orm.entity_manager')->refresh($menu);
-                }
-            }
-
             foreach ($post->getCollections() as $collection) {
                 $collection->setLocale($dataLocale);
                 $this->get('doctrine.orm.entity_manager')->refresh($collection);
             }
 
-
             $this->get('doctrine.orm.entity_manager')->refresh($post);
+        }
+        if ($transition = $configuration->getRequest()->request->get('transition')) {
+            try {
+                $this->get('workflow.registry')->get($post)
+                    ->apply($post, $transition);
+            } catch (ExceptionInterface $e) {
+            }
         }
 
         return $post;
+    }
+
+    /**
+     * @param Request $request
+     * @return RedirectResponse
+     */
+    public function applyTransitionAction(Request $request)
+    {
+        $configuration = $this->requestConfigurationFactory->create($this->metadata, $request);
+
+        try {
+            $resource = $this->findOr404($configuration);
+
+            $this->get('doctrine')->getManager()->flush();
+        } catch (ExceptionInterface $e) {
+            $this->get('session')->getFlashBag()->add('danger', $e->getMessage());
+            $resource = null;
+        }
+
+        return $this->redirectHandler->redirectToResource($configuration, $resource);
+    }
+
+    public function iframeAction(Post $post)
+    {
+        return $this->render('@PositibeCms/Page/_iframe.html.twig', ['content' => $post]);
     }
 } 
