@@ -10,13 +10,8 @@
 
 namespace Positibe\Bundle\NewsBundle\Entity;
 
-use Doctrine\ORM\Query;
-use Doctrine\ORM\QueryBuilder;
-use Gedmo\Translatable\TranslatableListener;
-use Positibe\Bundle\CmfBundle\Repository\LocaleRepositoryTrait;
-use Positibe\Bundle\OrmContentBundle\Entity\Traits\PageRepositoryTrait;
-use Positibe\Bundle\OrmRoutingBundle\Entity\HasRoutesRepositoryInterface;
-use Sylius\Bundle\ResourceBundle\Doctrine\ORM\EntityRepository;
+use Positibe\Bundle\CoreBundle\Repository\EntityRepository;
+use Positibe\Bundle\CoreBundle\Repository\LocaleRepositoryTrait;
 
 
 /**
@@ -25,19 +20,19 @@ use Sylius\Bundle\ResourceBundle\Doctrine\ORM\EntityRepository;
  *
  * @author Pedro Carlos Abreu <pcabreus@gmail.com>
  */
-class PostRepository extends EntityRepository implements HasRoutesRepositoryInterface
+class PostRepository extends EntityRepository
 {
     use LocaleRepositoryTrait;
 
     public function findOneByRoutes($route)
     {
         $qb = $this->createQueryBuilder('o')
-          ->addSelect('seo', 'image', 'r')
-          ->leftJoin('o.image', 'image')
-          ->leftJoin('o.seoMetadata', 'seo')
-          ->join('o.routes', 'r')
-          ->where('r = :route')
-          ->setParameter('route', $route);
+            ->addSelect('seo', 'image', 'r')
+            ->leftJoin('o.image', 'image')
+            ->leftJoin('o.seoMetadata', 'seo')
+            ->join('o.routes', 'r')
+            ->where('r = :route')
+            ->setParameter('route', $route);
 
         return $this->getQuery($qb)->getOneOrNullResult();
     }
@@ -45,13 +40,15 @@ class PostRepository extends EntityRepository implements HasRoutesRepositoryInte
     public function findLastNews($count)
     {
         $qb = $this->createQueryBuilder('o')
-          ->addSelect('image', 'routes')
-          ->leftJoin('o.image', 'image')
-          ->leftJoin('o.routes', 'routes')
-          ->setMaxResults($count)->orderBy(
-            'o.publishStartDate',
-            'DESC'
-          );
+            ->addSelect('image', 'routes')
+            ->leftJoin('o.image', 'image')
+            ->leftJoin('o.routes', 'routes')
+            ->where('o.state = :state')
+            ->setMaxResults($count)->orderBy(
+                'o.publishStartDate',
+                'DESC'
+            )
+            ->setParameter('state', 'published');
 
         return $this->getQuery($qb)->getResult();
     }
