@@ -13,6 +13,8 @@ namespace Positibe\Bundle\NewsBundle\Factory;
 use Doctrine\ORM\EntityManager;
 use Positibe\Bundle\NewsBundle\Entity\Comment;
 use Sylius\Component\Resource\Factory\Factory;
+use Sylius\Component\Resource\Factory\FactoryInterface;
+use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 
 
@@ -22,32 +24,49 @@ use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
  *
  * @author Pedro Carlos Abreu <pcabreus@gmail.com>
  */
-class CommentFactory extends Factory
+class CommentFactory implements FactoryInterface
 {
     protected $entityManager;
+    protected $defaultState;
 
-    public function __construct($className, EntityManager $entityManager)
+    public function __construct(EntityManager $entityManager)
     {
-        parent::__construct($className);
         $this->entityManager = $entityManager;
+    }
+
+    /**
+     * @return Comment
+     */
+    public function createNew()
+    {
+        return new Comment();
     }
 
     public function createByPostName($postName)
     {
-        if (!$post = $this->entityManager->getRepository('PositibeNewsBundle:Post')->findOneBy(
-          array(
-            'name' => $postName
-          )
-        )
-        ) {
-            throw new NotFoundHttpException('No se encontró la entrada con nombre ' . $postName);
+        if (!$post = $this->entityManager->getRepository('PositibeNewsBundle:Post')->findOneBy(['name' => $postName])) {
+            throw new NotFoundHttpException('No se encontró la entrada con nombre '.$postName);
         }
-        /** @var Comment $comment */
-        $comment = parent::createNew();
+
+        if (!$post->isCommentsEnabled()) {
+            throw new AccessDeniedHttpException('La entrada no es comentable');
+        }
+
+        $comment = $this->createNew();
 
         $comment->setPost($post);
+        $comment->setState($this->defaultState);
 
         return $comment;
 
     }
+
+    /**
+     * @param mixed $defaultState
+     */
+    public function setDefaultState($defaultState)
+    {
+        $this->defaultState = $defaultState;
+    }
+
 } 
