@@ -6,8 +6,9 @@
 namespace Positibe\Bundle\NewsBundle\Form\Type;
 
 use Symfony\Component\Form\AbstractType;
+use Symfony\Component\Form\Extension\Core\Type\ChoiceType;
 use Symfony\Component\Form\FormBuilderInterface;
-use Symfony\Component\OptionsResolver\OptionsResolverInterface;
+use Symfony\Component\OptionsResolver\OptionsResolver;
 
 /**
  * Class TagFormType
@@ -15,7 +16,6 @@ use Symfony\Component\OptionsResolver\OptionsResolverInterface;
  */
 class TagFormType extends AbstractType
 {
-
     private $locales;
 
     public function __construct($locales)
@@ -30,46 +30,41 @@ class TagFormType extends AbstractType
     public function buildForm(FormBuilderInterface $builder, array $options)
     {
         $builder
-          ->add(
-            'name',
-            null,
-            array(
-              'label' => 'Nombre'
-            )
-          )->add(
-            'locale',
-            'choice',
-            array(
-              'label' => 'Idioma',
-              'choices' => array_combine($this->locales, $this->locales)
-            )
-          )/*
-          ->add(
-            'enabled',
-            null,
-            array(
-              'label' => 'Habilitada'
-            )
-          )*/
-        ;
+            ->add(
+                'name',
+                null,
+                array(
+                    'label' => 'post_tag.form.name_label',
+                )
+            );
+        if ($options['data'] !== null && $options['data']->getId()) {
+            $builder->add(
+                'locale',
+                ChoiceType::class,
+                array(
+                    'label' => 'post_tag.form.locale_label',
+                    'choices' => array_combine($this->locales, $this->locales),
+                )
+            );
+        };
     }
 
     /**
-     * @param OptionsResolverInterface $resolver
+     * @param OptionsResolver $resolver
      */
-    public function setDefaultOptions(OptionsResolverInterface $resolver)
+    public function configureOptions(OptionsResolver $resolver)
     {
         $resolver->setDefaults(
-          array(
-            'data_class' => 'Positibe\Bundle\NewsBundle\Entity\Tag'
-          )
+            array(
+                'data_class' => 'Positibe\Bundle\NewsBundle\Entity\Tag',
+            )
         );
     }
 
     /**
      * @return string
      */
-    public function getName()
+    public function getBlockPrefix()
     {
         return 'positibe_post_tag';
     }
